Merged in hybrid-repo (pull request #57)
This PR sets up the structures and tooling for mvom to become a hybrid TS/JS repo.

The existing js code has been migrated to ./srcjs with the expectation that new and migrated code will be located in ./src.

A TS “Hello World” program has been included temporarily to ensure that linting and compilation is working as expected.