Merged in convert-errors (pull request #58)
This PR migrates the errors and constants from JS to TS.

Some deficiencies in the original approach for transpiling two separate folders into one were discovered – namely the babel-plugin-module-resolver did not properly convert any imports that crossed the folder boundary during transpilation. To resolve this, everything was moved into a single src folder with ts and js libraries kept separately. The build process looks to work now.

When the conversion is complete, the JS folder should be empty. That folder can then be deleted and the TS folder can be rolled into the src root.