Merged in support-clearing-arrays (pull request #60)
This PR fixes several issues with simple arrays and document arrays. It was discovered that MVOM could not clear out simple arrays and document arrays using the empty array syntax. It also contained another error where it was not possible to shorten an array because updates were being applied position-ally. If the array in the database contained 3 items, it was not possible to pass in an updated array of just two items. MVOM would update the first two items and leave the third item unmodified. The update includes first clearing the existing array before applying the update.

Type definitions for the Error classes were updated to include the Error suffix.

This is a breaking change.

I also added the `dist` folder to be ignored by Jest to silence this warning:

jest-haste-map: Haste module naming collision: mvom
  The following files share their name; please adjust your hasteImpl:
    * <rootDir>\package.json
    * <rootDir>\dist\package.json

Approved-by: Shawn McKnight <smcknight@storis.com>