Merged in CXM-2443-save-action-handle-schema-less-files (pull request #65)
This PR updates the compileModel save action in mvom. It adds a new clearAttributes flag to the options passed to executeDbFeature when executing the save action. The flag gets set to a boolean value based on whether the model's schema is null or not. If it is null, the flag will be set to true and the code in the save unibasic template will clear the record before writing anything to it. This was done because when we have a null schema, there are no specific attributes in the record, so it is essentially an array of attribute values of unknown length. This can lead to issues when trying to remove data from the end of the record because mvom only writes to the attributes it receives data for, so unless it specifically sets the values to null, it won’t remove them.

Approved-by: Shawn McKnight