Merged in fix-deployment (pull request #72)
This PR fixes an issue introduced in [CXM-2421](https://bitbucket.org/storis/mvom/pull-requests/66/cxm-2421-mvom-common/diff) where the deployment feature fails due to the `deployment` feature requiring the `setup` and `teardown` features to be installed. With this update these two features are considered to be "bootstrap" features along with the `deployment` feature and are therefore installed using the special stripped down "deploy" action before the rest of the features are deployed using the actual `deployment` feature.

Approved-by: Shawn McKnight <smcknight@storis.com>