Merged in CXM-2553-support-storis-encryption (pull request #77)
This change adds support to encrypt and decrypt sensitive data.

Support for a new property called `encrypted`, which defaults to false, has been added for string and ISO calendar date types. If this property is set to true, mvom expects the caller to have provided `encrypt` and `decrypt` functions. These will be used to encrypt the value before writing to the database after any transform and validation has occurred and decrypt the value after reading from the database and before any transformation has occurred.

Only ISO calendar date type support was added in addition to strings because STORIS encrypts date of birth.

The encrypt and decrypt functions must be overloaded so that they can accept either a single string or null value or an array or strings or null values. TS types have been updated. This was done so that the encryption and decryption could be handled exclusively by the `SimpleType`. The `ArrayType` uses the `SimpleType's` `getFromMVData` and `setIntoMVData` methods so the functions must be able to handle both single values and arrays.

Approved-by: Shawn McKnight <smcknight@storis.com>