Merged in fix-array-cast (pull request #81)
This is an addendum to PR #79 which stopped validating arrays as being required when the content schema type indicated it was required. The code from #79 works fine if an empty array is provided in the data, but if the property is not defined it still causes problems. This is because during validation the undefined value is converted to null in Document.validate. When the validation for the array type runs, it uses lodash castArray to turn the property into an array, which results in [null] and fails the schema type’s validation for required.

The PR adds a cast override to the BasePrimitiveArrayType which will cast null or undefined values to []. The current implementation uses the BaseType's implementation which just leaves the passed value unmutated. This is similar to how DocumentArrayType currently works.

Approved-by: Ken Thompson