Merged in CXM-3747-node12-support (pull request #90)
This PR updates mvom to make node 12.22 the minimum required version. This is the latest node 12 release. The minor version has been bumped to 0.10.0 to indicate a breaking change.

With this change the use of lodash flatten is no longer needed as Array.prototype.flat is now available natively. I also removed the babel class properties plugin as public class properties are natively available in node 12. Finally, babel-register was also removed, which may have been a peer dependency of babel-node at some point.

Approved-by: Shawn McKnight