Merged in CXM-3931-u2-foreign-key (pull request #94)
This PR enhances mvom to perform foreign key validations on the U2 server. Foreign key definitions can be defined directly in the schema and will be executed during save. The file will first be read and locked and then fk validations will be performed serially using readv. By moving the foreign key validation to the server we greatly reduce the number of pooled processes required during save and reduce load on the u2 server and our services since we no longer need to read and pass back large records.

As part of this task I improved the Typescript types for the mvom schema. The string type can now accept a foreign key definition in one of two forms:

```
interface SchemaForeignKeyDefinition {
  file: string | string[];
  keysToIgnore?: string[];
  entityName: string;
}

interface PositionForeignKeyDefinition {
  [key: number]: SchemaForeignKeyDefinition;
}

type SchemaCompoundForeignKeyDefinition = PositionForeignKeyDefinition & {
  splitCharacter: string;
};
```

In addition, a new option called idForeignKey has been added to the Schema options so the same validations can be run against the record id.

This support all current fk requirements including.\
1\. validating one attribute against multiple files (ORDER/billTo → CUSTOMER/WAREHOUSE.LOCATION)\
2\. ignoring specific keys ie ignoring DELETED in the order header for deleted lines when validating against the PRODUCT file.

3\. Compound ids that need to be split.

The routine de-duplicates files and entityIds.

I added a new ForeignKeyValidationError that will be omitted with all foreign key validation errors so the mvom adapter in moleculer services will need to be enhanced to log and report this.

Approved-by: Shawn McKnight