Convert Schema, Document, and schemaType from JS to TS (#14)
This PR converts the Schema class, the Document class, and the various schemaType classes from JavaScript to TypeScript. While working on this conversion, several changes were made to the codebase for one or more of the following reasons:

The TypeScript compiler alerted to errors that were previous undetected and were fixed as part of the change
The TypeScript compiler alerted to codepaths that were not really accessible without violating type rules and the codepaths were removed or altered accordingly
The TypeScript compiler alerted to inconsistencies in API structures in subclasses
I will attempt to note as many of the changes as I can inline at the point of the changes.

During the conversion, all existing tests were scrapped and rewritten from scratch. I believe the tests are much more solid now and exercise the library much better than before.

Additionally, some cleanup was performed to ensure all tests are in __tests__ folders rather than as siblings of the unit under test.