Convert Connection, Query, and compileModel to TS (#18)
This PR converts the Connection and Query classes as well as the compileModel function from JavaScript to TypeScript. Other than the entrypoint index, this completed the conversion of all the packaged JavaScript (some unpackaged script files still remain and will be converted later).

Since there is basically no JavaScript remaining and all existing tests have been rewritten, much of the existing dev tool configuration was able to be streamlined. babel-plugin-rewire-ts was eliminated from the babel configuration and the configuration was updated to handle the existing environment and file structures. Some eslint overrides for javascript were eliminated. Similar changes were made to the jest config. Coverage requirements were increased back to 100%. Unused dependencies, including sinon were dropped.

This work allowed the input and output from the Db server subroutines to be typed which is a huge improvement in interacting with that layer.

During the conversion, some bugs were uncovered and corrected and some code was refactored to improve its structure. I will attempt to highlight those changes in comments inline with the code.