Final conversion of repo from JS to TS (#20)
This PR completes the conversion of the repo from JavaScript to TypeScript. Included in this PR:

Remove files from nested src/ts folder to the root src folder
Eliminate the #shared alias from tsconfig, babel config, and eslint config. This is because declaration file generation does not work with the aliased paths, and the benefit of that outweighs the benefit of the aliasing.
Elimination of the src/shared folder in favor of src/group folders.
Add a declaration file generation step to the build process
Change the pointer to the main index and the declaration file in package.json
Convert build scripts from js to ts and refactor to improve their structures
Use js private field and method syntax in favor of ts private directive to ensure that delcaration file could be built for Model and its parent Document. TS will not build the declaration file for members declared as private which are part of a class returned from a function.
Allow skip/limit in Query constructor to accept null
Update the main index to export useful types
Although there are a lot of files in this PR, most of it is renames along with changes to use relative pathing.