Miscellaneous Refactorings (#22)
This PR contains several refactoring efforts in order to improve the overall code structure. It will be the last pull request to finalize the conversion from TypeScript to JavaScript. Changes included in this PR are as follows:

Any multi-line jsdocs which did not need to be multi-line were changed to the single-line format to improve readability
The BaseError class was properly defined as abstract
The ConnectionManagerError class was renamed as MvisError
Interfaces named ConstructorOptions were prefixed with the class they are used in constructing
Several uses of GenericObject (i.e. Record<string, any>) were modified to be a more concrete type. Typically this was for scenarios where a more concrete type already existed but the type was being misused.
The readonly property modifier was added to as many places as possible since generally speaking most class properties should be readonly
The getFromMvArray "utility" was migrated to be a class method of BaseSchemaType instead. After correcting the parameter type to be MvRecord, a type issue/bug was found which required the addition of an Array.isArray check on the data to narrow the type.
The handleRequiredValidation utility's logic was migrated to the createRequiredValidator method in BaseScalarType. As part of this change, that logic no longer checks the required flag itself, but instead now relies on the created validator.
The handleTypeValidation utility's logic was migrated to the createTypeValidator method in BaseScalarType. A new validateType method was added along with this which always returns true and is expected to be overridden in subclasses which do type validation (e.g. number, date types).
Type validation now always runs on any subclass of BaseScalarType and will either run the superclass' built-in variant or the overridden one as appropriate
Types were relocated to be closer to the class that uses them as appropriate. In scenarios where there wasn't a good class to associate the types with, new type files were created to make the index of the types folder simply a re-export point.
Validation functions which did not need to be asynchronous were changed to be synchronous
The paths property of Schema was changed from an object to a Map. Additionally, the method which constructs those paths was changed to return the calculated paths to be set by the constructor instead of the method mutating the instance member.
The dictPaths property of Schema was changed from an object to a Map
The returned value from Document.validate() was changed from an object to a Map
The mockLogger test helper was deleted as it was not being used
Rename #__id to #_id in compileModel