Allow type information for dictionaries (#53)
This PR introduces a new feature to allow for specification of type information for additional dictionaries provided as options to the Schema constructor. Currently, dictionaries can be added through these options to allow queries to be executed against dictionaries not defined in the schema, such as virtual dictionaries. However, the constants provided for these dictionaries cannot be transformed into multivalue format because there is no type information. For instance, if a dictionary resolves to a multivalue boolean (1 or 0) then the consumer must provide a 1 or a 0 as the query conditional's constant value rather than the more familiar true or false. With the addition of type information, the query constants will be transformed to multivalue internal format in the same manner as dictionaries defined in the schema definition.

In order to allow for backwards compatibility, the existing format of providing dictionaries as key/value pairs representing a property and a dictionary was retained. When using this syntax, the constants will be treated as strings and the only casting done will be to convert any input value to a string. This should match the existing behavior. It is now additional possible to provide an object in place of the dictionary name and supply a dictionary property (which will contain the dictionary ID) and a type property (which will match the type names of the schema definition scalars). Additionally, the ISOCalendarDateTime and ISOTime types will allow for a dbFormat to be specified to indicate if the time is in seconds or milliseconds, similar to the schema type option.

The implementation of this change involved the introduction of a new DataTransformer interface. The data transformation logic resident in each schema type was then moved to a class implementing this interface which only facilitates the data transformation logic and does not offer the other logic resident in the schema type. The scalar schema types were adjusted to also implement this same interface, and the existing transformation methods now delegate to the corresponding DataTransformer class implementation. When adding dictionaries with type information, an appropriate DataTransformer will be instantiated and used for data transformations for those. Dictionaries defined by schema types will continue to use the transformation methods in the schema types, but as previously stated those merely delegate to another DataTransformer implementation. This allows for both schema types and data transformers to perform the transformations with a common interface.