Allow user defined properties in Model methods (#76)
This PR allows for custom user defined properties to be passed to the Model methods as a property of an object named userDefined. This object can contain up to 5 properties named option1 , ..., option5. These properties will be passed to the database subroutines and added to the existing S$MVOM named common block as S$MVOM.USER1, ..., S$MVOM.USER5.

These common variables then can be used in upstream functions such as database triggers.

One additional change was made as well with the .mvomrc.json file being renamed as manifest.json which is a more appropriate term to use for this file.