Validate server limits for queries (#79)
This PR adjusts the query process to validate queries prior to execution to ensure they do not exceed the limits established by the database server. The getServerInfo database feature will execute the LIMIT verb, parse the output, and return the maximum sentence length, with conditions, and sort conditions which can be in a query. These values will be used to compare against the query that is to be executed, and if there is a violation of the server limits then a new QueryLimitError will be thrown and the query will not be executed.