Allow projection of only ids (#86)
Currently, the behavior of the projection option (find, findAndCount, findById, findByIds) is such that if the consumer were to provide an empty array of projection values then the behavior would be identical to providing no projection option at all. Only if there are one or more values in the projection array will projection change the output result

This behavior is unexpected as an empty projection array should indicate that nothing should be projected at all. This change adjusts the behavior so that no data will be projected if an empty array is provided for projection. The _id and __v values will continue to be emitted from the db server, however, and will be valued in the resulting document.