Change Unibasic deployment to use MVIS Admin (#186)
Currently, MVOM expects a bootstrapper program to be defined in the MVIS REST subroutine definitions named mvom_entry@n.n.n where n.n.n represents the version of the entry process. This subroutine definition must exist prior to MVOM opening a connection which adds a lot of friction to the installation and configuration process. MVOM then uses that bootstrapper to deploy other database subroutines which are then utilized by MVOM for its functionality.

This PR completely changes how the deployment process works to instead rely on MVIS Admin's API to generate the REST subroutine definitions and deploy server side code. When establishing a connection, consumers must now specify the URL of MVIS Admin as well as a valid admin username and password. MVOM will then use MVIS Admin's API to determine if the REST subroutine definition is available prior to opening the connection. In line with existing behavior, if this REST subroutine does not exist, an error will be thrown. Consumers can then choose to deploy the db server functionality which will generate a REST subroutine definition, deploy the BASIC source code to the db server, and compile and catalog it. This functionality is located in the DeploymentManager class, but deployment will still be handled by consumers on the Connection instance.

Previously, MVOM's BASIC code would consist of a single entry routine and several associated external subroutines. These subroutines all used semantic versioning based on an included manifest.json file so that it would only require deployment if the versions were incompatible based on semantic versioning. This has changed substantially with this PR. Now MVOM will consist solely of a single subroutine named mvom_main which includes internal callable subroutines for each separate feature. These subroutines continue to be templated and built using nunjucks templating engine to allow for the BASIC source to remain compartmentalized for maintainability purposes. At build time, the mvom_main program will be suffixed with an @ sign followed by a 8 character hash of the source code. This will effectively allow for the bundled server side source code to be versioned via the hash identifier. Unlike previously where semantic versioning may have eliminated the need to deploy versions of the features, now each version of MVOM will require a specific server side bundle to be associated with it. This mechanism greatly simplifies the deployment process and reduces the risk of unintentional errors in the semantic versioning syntax. Essentially, any time server side code changes, a new bundle will be generated for deployment and will be the version that is used by that version of MVOM.

With the changes to the basic code, the unibasicTemplates folder was reorganized into constants and subroutines folders. subroutines was further separated into external and internal folders with external representing subroutines that will be called from MVOM and internal representing subroutines which are called by other subroutines. This convention should be easier to understand go forward.

As part of this change, logging became necessary in classes that did not have access to a Connection instance. Previously, all logging was handled on the Connection, but the need to have logging outside of a connection was apparent. Therefore, logging was abstracted into a new LogHandler class. This allowed logging to be performed in any class that was provided a LogHandler instance. Creating a Connection will instantiate a new LogHandler which is provided with the Logger passed to the Connection constructor and that instance will be passed to all classes requiring logging.