Return write error status codes (#322)
This change improves the reporting of write errors by returning additional error codes that map to the result of the `status()` function, which is now called in the `write` error clause.

The additional error codes can now distinguish that a write failed due to a "system" error ie a permission error or file corruption, violating a trigger constraint ie the update trigger rejected the write, there was an error with the trigger, or a no duplicate index constraint was violated. An "unknown write error" code was also created in case an unknown status is returned. All four `status()` return values defined in the UniBasic reference guide were mapped.

A note about the "trigger error" return code. I was able to replicate this by decataloging the trigger program. If the trigger itself aborts then this status code is not returned and UniObjects returns its own subroutine failed error code.