Allow specification of a maximum return payload size from the database server (#357)
This PR adds a new configurable option for a maximum allowed return
payload size. Either due to bad data or an incorrectly configured query
the UniRPC server may return an extremely large payload which can
saturate a network or cause MVIS to become unresponsive. If the payload
size exceeds the maximum, a DbServerError will be thrown.

I chose a high default value of 100MB but made it a configurable
connection and individual request option.