CXM-5384 Fix fk validation with arrays of file names (#467)
It was recently discovered that even though we can specify an array of
file names in the `foreignKey` property of our mvom schemas to validate
the foreign key against, it is only ever validating against the first
file. This PR changes the file name/entity name separator character from
a `,` to a `#` since when an array of file names gets transformed to a
string it inserts commas between each of them. Thus, when performing
splitting logic later on in the process flow, it was not working
correctly. After making this change the file names, were correctly split
from the entity name, but additional splitting was needed to split each
of the file names back into an array to pass to the u2 subroutine.

---------

Co-authored-by: Ken Thompson <kat@storis.com>