Eliminate @storis dependencies internalize configs (#644)
This PR removes the two storis dependencies `@storis/eslint-config` and
`@storis/tsconfig`. Those repositories are no longer being maintained,
so their configs will be internalized.

This PR also updates the `eslint-config-prettier` plugin.

There was an attempt made to upgrade eslint to v9 as well, but there are
plugins that are yet to support v9 of eslint. They have been documented
in a tracker issue - [#646](https://github.com/STORIS/mvom/issues/646)