Fix Projection (#653)
This PR fixes the format_document u2 subroutine's handling of the
`projection` option.
Before this fix, when formatting the output document it was not using
the passed in `projection`. It was using the loop index. This PR ensures
the provided projection positions get used when formatting the output
document.

Tested this out by creating a schema that covers the different schema
types and using projections on those types. Scalar, scalar array,
document array, object schema definition, and use of `new Schema`.
Looked good.